refactor: 코드 리팩토링 및 application.yml 추가
