+++ src/main/java/com/sproutt/eussyaeussyachat/application/chat/ChatServiceImpl.java
@@ -1,9 +1,8 @@
 package com.sproutt.eussyaeussyachat.application.chat;
 
-import com.sproutt.eussyaeussyachat.application.chat.ChatService;
+import com.sproutt.eussyaeussyachat.api.dto.OneToOneChatMessageDTO;
 import com.sproutt.eussyaeussyachat.domain.ChatMessageRepository;
 import com.sproutt.eussyaeussyachat.domain.OneToOneChatMessage;
-import com.sproutt.eussyaeussyachat.api.dto.OneToOneChatMessageDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
