+++ src/main/java/com/sproutt/eussyaeussyachat/config/RedisServerConfig.java
@@ -1,26 +1,35 @@
 package com.sproutt.eussyaeussyachat.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sproutt.eussyaeussyachat.application.redisServer.RedisSubscriber;
 import com.sproutt.eussyaeussyachat.domain.OneToOneChatMessage;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.data.redis.listener.RedisMessageListenerContainer;
 import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
+@RequiredArgsConstructor
 public class RedisServerConfig {
+
+    @Value("${redis.server.host}")
+    private String host;
+
+    @Value("${redis.server.port}")
+    private int port;
 
     @Bean(name = "redisServerConnectionFactory")
     @Primary
     public LettuceConnectionFactory redisServerConnectionFactory() {
-        return new LettuceConnectionFactory("127.0.0.1", 6380);
+        return new LettuceConnectionFactory(host, port);
     }
 
     @Bean
@@ -28,6 +37,11 @@ public class RedisServerConfig {
         RedisMessageListenerContainer container = new RedisMessageListenerContainer();
         container.setConnectionFactory(connectionFactory);
         return container;
+    }
+
+    @Bean
+    public ChannelTopic channelTopic() {
+        return new ChannelTopic("Server_A");
     }
 
     @Bean(name = "redisServerTemplate")
